Moved sort function in the modifier bar
