+++ src/ui/summary-ui-handler.ts
@@ -37,7 +37,8 @@ import { PlayerGender } from "#enums/player-gender";
 import { Stat, PERMANENT_STATS, getStatKey } from "#enums/stat";
 import { Nature } from "#enums/nature";
 import { achvs } from "#app/system/achv";
-import { modifierSortFunc, PokemonHeldItemModifier } from "#app/modifier/held-item-modifier";
+import { PokemonHeldItemModifier } from "#app/modifier/held-item-modifier";
+import { modifierSortFunc } from "#app/modifier/modifier-bar";
 
 enum Page {
   PROFILE,
